# ice-bubble-admin/src/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ice-bubble Admin API
  description: ice-bubble 管理后台 REST API 文档
  version: 1.4.1
  contact:
    name: ice-bubble team
servers:
  - url: http://localhost:13000
    description: 本地开发
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer Token 认证（配置在 auth.token 或 ICE_AUTH_TOKEN）
  schemas:
...
    Module:
      type: object
      properties:
        moduleKey:
          type: string
        name:
          type: string
        baseUrl:
          type: string
        enabled:
          type: boolean
        pollInterval:
          type: integer
        registeredTime:
          type: string
        version:
          type: string
        status:
          type: object
          properties:
            state:
              type: string
              enum: [running, stopped, error]
            lastPollTime:
              type: string
              nullable: true
            lastError:
              type: string
              nullable: true
            latencyMs:
              type: integer
              nullable: true
            runtime:
              type: object
              nullable: true
              properties:
                startTime:
                  type: string
                  nullable: true
                uptimeSeconds:
                  type: integer
                  nullable: true
                messagesCollected:
                  type: integer
                  nullable: true
                errorsCount:
                  type: integer
                  nullable: true
...
paths:
...
  /api/modules:
    get:
      tags: [Modules]
      summary: 模块列表
      responses:
        '200':
          description: 模块列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  modules:
                    type: array
                    items:
                      $ref: '#/components/schemas/Module'
    post:
      tags: [Modules]
      summary: 新增模块
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [moduleKey, name, baseUrl]
              properties:
                moduleKey:
                  type: string
                name:
                  type: string
                baseUrl:
                  type: string
                enabled:
                  type: boolean
                  default: true
                pollInterval:
                  type: integer
                  default: 30000
      responses:
        '201':
          description: 模块创建成功
        '400':
          description: 参数错误
        '409':
          description: 模块已存在
  /api/modules/test-connection:
    post:
      tags: [Modules]
      summary: 测试模块连接
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [baseUrl]
              properties:
                baseUrl:
                  type: string
      responses:
        '200':
          description: 连接成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  moduleKey:
                    type: string
                  moduleType:
                    type: string
                  status:
                    type: string
                  version:
                    type: string
  /api/modules/{key}:
    get:
      tags: [Modules]
      summary: 获取单个模块
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 模块详情
        '404':
          description: 模块不存在
    put:
      tags: [Modules]
      summary: 更新模块
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                baseUrl:
                  type: string
                enabled:
                  type: boolean
                pollInterval:
                  type: integer
      responses:
        '200':
          description: 更新成功
        '403':
          description: 禁止更新 admin
        '404':
          description: 模块不存在
    delete:
      tags: [Modules]
      summary: 删除模块
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 删除成功
        '403':
          description: 禁止删除 admin
        '404':
          description: 模块不存在
  /api/modules/{key}/status:
    get:
      tags: [Modules]
      summary: 手动触发模块状态更新
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 模块状态
        '404':
          description: 获取状态失败
  /api/modules/{key}/config:
    get:
      tags: [Modules]
      summary: 获取模块运行时配置
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 模块配置
        '404':
          description: 模块不存在

  # ── Media ──
...